JIRA-ID:[ODUHIGH-295]- Tunnel creation in EGTP

+++ src/cm/common_def.c
@@ -92,7 +92,7 @@ void freqDomRscAllocType0(uint16_t startPrb, uint16_t prbSize, uint8_t *freqDoma
  * ****************************************************************/
 void oduCpyFixBufToMsg(uint8_t *fixBuf, Buffer *mBuf, uint16_t len)                            
 {
-   uint8_t idx, revIdx, temp;
+   uint16_t idx, revIdx, temp;
    uint16_t copyLen;
 
    /* ODU_COPY_FIX_BUF_TO_MSG copies fixed buffer in reverse order. \
